do basic setup
