+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "tdec-register",
+  "version": "1.0.0",
+  "description": "College Leave Applications",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/TDEC-surat/TDEC-Register.git"
+  },
+  "author": "Dharmik, Aksh, Meet",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/TDEC-surat/TDEC-Register/issues"
+  },
+  "homepage": "https://github.com/TDEC-surat/TDEC-Register#readme",
+  "dependencies": {
+    "ejs": "^3.1.6",
+    "ejs-mate": "^3.0.0",
+    "express": "^4.17.3"
+  }
+}
