+++ views/dashboard.ejs
@@ -1,0 +1,2 @@
+<% layout('layouts/boilerplate') -%>
+<h1>DASHBOARD</h1>
