+++ views/layouts/boilerplate.ejs
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>TDEC</title>
+	</head>
+	<body>
+		<%- body -%>
+	</body>
+</html>
